requirements.txt
simplejson
watchdog
Flask
pylint
flask-bcrypt
flask-login
Flask-WTF
python-dotenv